LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-close-via-attribute.tentative-expected.txt

button

FAIL Removing the open attribute from an open modal dialog should run the closing algorithm. assert_equals: Previously focused element should be focused after waiting for a task. expected Element node <button>button</button> but got Element node <button>button in dialog</button>
FAIL Removing the open attribute from an open non-modal dialog should fire a close event. assert_equals: Previously focused element should be focused after waiting for a task. expected Element node <button>button</button> but got Element node <body><button>button</button>
<dialog>
  <button>button i...
